cleaned up a couple odd type choices

+++ examples/gaussian-convolution.rs
@@ -11,16 +11,16 @@ fn main() {
     let x_max = 10.0;
     let n_steps = 10000;
     for i in 0..n_steps {
-        let x_value = (x_max - x_min) * i as f32 / n_steps as f32 + x_min;
+        let x_value = (x_max - x_min) * i as f64 / n_steps as f64 + x_min;
         let f = |x: f64| {
             if x != 0.0 {
-                x.sin() / x * gaussian(x_value as f64 - x)
+                x.sin() / x * gaussian(x_value - x)
             } else {
-                1.0 * gaussian(x_value as f64 - x)
+                1.0 * gaussian(x_value - x)
             }
         };
 
-        conv.push((x_value, integrator.integrate(f, -100.0, 100.0, n_steps) as f32));
+        conv.push((x_value, integrator.integrate(f, -100.0, 100.0, n_steps) as f64));
     }
 
     plot_result(conv).unwrap();
@@ -31,7 +31,7 @@ fn gaussian(x: f64) -> f64 {
     (-1.0 * x * x).exp()
 }
 
-fn plot_result(data: Vec<(f32, f32)>) -> Result<(), Box<dyn std::error::Error>> {
+fn plot_result(data: Vec<(f64, f64)>) -> Result<(), Box<dyn std::error::Error>> {
     let root = BitMapBackend::new("gaussian-convolution-example.png", (640, 480)).into_drawing_area();
     root.fill(&WHITE)?;
 
@@ -40,7 +40,7 @@ fn plot_result(data: Vec<(f32, f32)>) -> Result<(), Box<dyn std::error::Error>> 
         .margin(5)
         .x_label_area_size(30)
         .y_label_area_size(30)
-        .build_cartesian_2d(-10f32..10f32, -2f32..2f32)?;
+        .build_cartesian_2d(-10f64..10f64, -2f64..2f64)?;
 
     chart.configure_mesh().draw()?;
